Avoid file clashes when the feed reuses a filename.
If the feed uses the same filename multiple times the downloaded files would
overwrite each other, since only the basename is used when writing the file.
Detect this case and if it happens rename the all downloaded files by using the
title instead of the filename used on the server.